refactor(ui): split the swipe row's geometry out of its component
CI's `detektMetadataCommonMain` is not the per-module `detekt` task, which
is why this passed locally: the file had reached the 15-function limit, and
the group's pointer loop had two ways out of it.

The pure functions move to `SwipeActionRowGeometry.kt` — the reveal, the
settle, the released hand-back, the displaced opacity, the commit
threshold. Nothing there springs, remembers or reads a theme, which is
what makes them the tested half; the component decides what the travel is,
and they decide what the row looks like at it.

The group's tap watcher leaves its loop one way now instead of two, which
is the same two conditions written as one.
